Revert "Enrollment"

+++ app/Services/AnnouncementService.php
@@ -3,6 +3,8 @@
 namespace App\Services;
 
 use App\Repositories\AnnouncementRepository;
+use App\Models\Course;
+
 use Carbon\Carbon;
 
 class AnnouncementService
@@ -16,7 +18,7 @@ class AnnouncementService
 
     public function create(array $data)
     {
-        // Ensure publishDate is set to now if not provided
+       
         $data['publishDate'] = $data['publishDate'] ?? Carbon::now();
 
         return $this->announcements->create([
@@ -24,7 +26,7 @@ class AnnouncementService
             'content'    => $data['content'],
             'category'   => $data['category'] ?? null,
             'publishDate'=> $data['publishDate'],
-            'courseID'   => $data['courseID'] ?? null, // optional
+            'courseID'   => $data['courseID'] ?? null,
         ]);
     }
      public function editAnnouncement(int $id, array $data)
@@ -32,21 +34,37 @@ class AnnouncementService
         $announcement = $this->announcements->findById($id);
 
         if (!$announcement) {
-            return null;
+            abort(404, 'Announcement not found.');
         }
-
+        $user = auth('api')->user();
+        $course = Course::find($announcement->courseID);
+        if (!$course || $course->adminid !== $user->id) {
+            abort(403, 'You are not authorized to edit this announcement.');
+        }
         $updateData = [
             'title'       => $data['title'] ?? $announcement->title,
             'content'     => $data['content'] ?? $announcement->content,
             'category'    => $data['category'] ?? $announcement->category,
             'publishDate' => $data['publishDate'] ?? Carbon::now(),
-            'courseID'    => $data['courseID'] ?? $announcement->courseID,
+            'courseID'    => $announcement->courseID,
         ];
 
         return $this->announcements->update($announcement, $updateData);
     }
     public function deleteAnnouncement(int $id): bool
     {
+        $announcement = $this->announcements->findById($id);
+
+        if (!$announcement) {
+            abort(404, 'Announcement not found.');
+        }
+
+        $user = auth('api')->user();
+        $course = Course::find($announcement->courseID);
+        if (!$course || $course->adminid !== $user->id) {
+           abort(403, 'You are not authorized to delete this announcement.');
+        }
+
         return $this->announcements->delete($id);
     }
     public function getallAnnouncements()
@@ -55,5 +73,10 @@ class AnnouncementService
     }
     public function getAnnouncement(int $id){
         return $this->announcements->findById($id);
+    }
+
+     public function getAnnouncementsByCourseId(int $courseId)
+    {
+        return $this->announcements->getByCourseId($courseId);
     }
 }
